// backend/src/validations/params.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by validationCompile.js.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run jsonSchema2ts.js to regenerate this file.
 */
export namespace Device {
    export interface Get {
        id?: string;
        shellyID?: string;
    }
    export interface Call {
        id?: number;
        shellyID: string;
        method: string;
        params?: {
            [k: string]: unknown;
        };
    }
    export interface GetSetup {
        shellyID: string;
        mode?: string;
    }
    export interface Delete {
        shellyID: string;
        id: number;
    }
}
export namespace FleetManager {
    export interface Subscribe {
        events: string[];
        options?: {
            [k: string]: unknown;
        };
    }
    export interface Unsubscribe {
        ids: number[];
    }
}
export namespace Grafana {
    export interface GetDashboard {
        slug: string;
    }
}
export namespace Group {
    export interface Create {
        name: string;
        devices?: string[];
    }
    export interface Add {
        id: string | number;
        shellyID: string;
    }
    export interface Remove {
        id: string | number;
        shellyID: string;
    }
    export interface Delete {
        id: string | number;
    }
    export interface List {
        id?: string | number;
    }
    export interface Get {
        id: string | number;
    }
    export interface Rename {
        id: string | number;
        newName: string;
    }
    export interface Find {
        shellyID: string;
    }
    export interface Set {
        id: string | number;
        name: string;
        devices: string[];
    }
}
export namespace User {
    export interface Create {
        username?: string;
        password: string;
        name: string;
        fullName: string;
        email: string;
        group: string;
        permissions: string[];
    }
    export interface Update {
        id: number;
        name?: string;
        fullName?: string;
        password?: string;
        group?: string;
        email?: string;
        permissions?: string[];
        enabled?: boolean;
    }
    export interface Delete {
        id: number;
    }
    export interface Authenticate {
        username: string;
        password: string;
        purpose?: 'alexa';
        endpoint?: string;
    }
    export interface Refresh {
        refresh_token: string;
    }
}
export namespace WaitingRoom {
    export interface AcceptPendingById {
        ids: number[];
    }
    export interface AcceptPendingByExternalId {
        externalIds: string[];
    }
    export interface RejectPending {
        shellyIDs: string[];
    }
}